backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-socketio==5.13.0
pydantic==2.11.7